requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
passlib
bcrypt
python-jose